Symmetric: Constant time compare hashes of encryption and decryption keys on decrypt. Refactor

+++ ferrocrypt-lib/src/symmetric.rs
@@ -5,7 +5,7 @@ use argon2::Variant;
 use chacha20poly1305::aead::Aead;
 use zeroize::Zeroize;
 use crate::{archiver, CryptoError};
-use crate::common::{get_file_as_byte_vec, get_file_stem_to_string, normalize_paths};
+use crate::common::{constant_time_compare_256_bit, get_file_as_byte_vec, get_file_stem_to_string, normalize_paths, sha3_hash};
 use crate::CryptoError::{ChaCha20Poly1305Error, Message};
 
 
@@ -14,12 +14,13 @@ mod tests {
     use std::fs;
     use crate::CryptoError;
     // use zeroize::Zeroize;
-    use crate::symmetric::{decrypt_file, encrypt_file};
+    use crate::symmetric::{decrypt_file, decrypt_large_file, encrypt_file, encrypt_large_file};
 
     const SRC_FILE_PATH: &str = "src/test_files/test-file.txt";
     const SRC_DIR_PATH: &str = "src/test_files/test-folder";
     const ENCRYPTED_FILE_PATH: &str = "src/dest/test-file.fcs";
     const ENCRYPTED_DIR_PATH: &str = "src/dest/test-folder.fcs";
+    const ENCRYPTED_LARGE_FILE_PATH: &str = "src/dest/test-file.fcls";
     const DEST_DIR_PATH: &str = "src/dest/";
     const PASSPHRASE: &str = "strong_passphrase";
 
@@ -38,7 +39,7 @@ mod tests {
     #[test]
     fn decrypt_file_test() -> Result<(), CryptoError> {
         // let mut password = rpassword::prompt_password("password:")?;
-        let mut passphrase = PASSPHRASE.to_string();
+        let mut passphrase = "strong_passphrase".to_string();
         decrypt_file(ENCRYPTED_FILE_PATH, DEST_DIR_PATH, &mut passphrase)?;
 
         // password.zeroize();
@@ -68,13 +69,36 @@ mod tests {
 
         Ok(())
     }
+
+    #[test]
+    fn encrypt_large_file_test() -> Result<(), CryptoError> {
+        fs::create_dir_all("src/dest")?;
+        // let mut passphrase = rpassword::prompt_password("passphrase:")?;
+        let mut passphrase = PASSPHRASE.to_string();
+        encrypt_large_file(SRC_FILE_PATH, DEST_DIR_PATH, &mut passphrase)?;
+
+        // passphrase.zeroize();
+
+        Ok(())
+    }
+
+    #[test]
+    fn decrypt_large_file_test() -> Result<(), CryptoError> {
+        // let mut password = rpassword::prompt_password("password:")?;
+        let mut passphrase = "strong_passphrase".to_string();
+        decrypt_large_file(ENCRYPTED_LARGE_FILE_PATH, DEST_DIR_PATH, &mut passphrase)?;
+
+        // password.zeroize();
+
+        Ok(())
+    }
 }
 
 // Encrypt file with XChaCha20Poly1305 algorithm
-pub fn encrypt_file(src_file_path: &str, dest_dir_path: &str, passphrase: &mut str) -> Result<(), CryptoError> {
-    let (src_file_path_norm, dest_dir_path_norm) = normalize_paths(src_file_path, dest_dir_path);
-    let file_stem = &archiver::archive(&src_file_path_norm, &dest_dir_path_norm)?;
-    let file_name_zipped = &format!("{dest_dir_path_norm}{file_stem}.zip");
+pub fn encrypt_file(input_path: &str, output_dir: &str, passphrase: &mut str) -> Result<(), CryptoError> {
+    let (input_path_norm, output_dir_norm) = normalize_paths(input_path, output_dir);
+    let file_stem = &archiver::archive(&input_path_norm, &output_dir_norm)?;
+    let file_name_zipped = &format!("{}{}.zip", output_dir_norm, file_stem);
     println!("\nencrypting {} ...", file_name_zipped);
 
     let argon2_config = argon2_config();
@@ -84,6 +108,10 @@ pub fn encrypt_file(src_file_path: &str, dest_dir_path: &str, passphrase: &mut s
     OsRng.fill_bytes(&mut nonce);
 
     let mut key = argon2::hash_raw(passphrase.as_bytes(), &salt, &argon2_config)?;
+
+    // Hash the encryption key for comparison when decrypting
+    let key_hash_ref: [u8; 32] = sha3_hash(&key)?;
+
     let cipher = XChaCha20Poly1305::new(key[..32].as_ref().into());
     let file_original = get_file_as_byte_vec(file_name_zipped)?;
     let ciphertext = cipher.encrypt(nonce.as_ref().into(), &*file_original)?;
@@ -91,16 +119,17 @@ pub fn encrypt_file(src_file_path: &str, dest_dir_path: &str, passphrase: &mut s
         .write(true)
         .append(true)
         .create_new(true)
-        .open(format!("{}{}.fcs", &dest_dir_path_norm, file_stem))?;
+        .open(format!("{}{}.fcs", &output_dir_norm, file_stem))?;
 
     file_path_encrypted.write_all(&salt)?;
     file_path_encrypted.write_all(&nonce)?;
+    file_path_encrypted.write_all(&key_hash_ref)?;
     file_path_encrypted.write_all(&ciphertext)?;
 
     fs::remove_file(file_name_zipped)?;
-    let file_name_encrypted = &format!("{dest_dir_path_norm}{file_stem}.fcs");
+    let file_name_encrypted = &format!("{}{}.fcs", output_dir_norm, file_stem);
     println!();
-    println!("encrypted to {file_name_encrypted}");
+    println!("encrypted to {}", file_name_encrypted);
 
     salt.zeroize();
     nonce.zeroize();
@@ -111,35 +140,46 @@ pub fn encrypt_file(src_file_path: &str, dest_dir_path: &str, passphrase: &mut s
 }
 
 // Decrypt file with XChaCha20Poly1305 algorithm
-pub fn decrypt_file(encrypted_file_path: &str, dest_dir_path: &str, passphrase: &mut str) -> Result<(), CryptoError> {
-    let (encrypted_file_path_norm, dest_dir_path_norm) = normalize_paths(encrypted_file_path, dest_dir_path);
+pub fn decrypt_file(input_path: &str, output_dir: &str, passphrase: &mut str) -> Result<(), CryptoError> {
+    let (input_path_norm, output_dir_norm) = normalize_paths(input_path, output_dir);
 
-    if encrypted_file_path_norm.ends_with(".fcs") {
-        println!("decrypting {} ...\n", encrypted_file_path);
+    if input_path_norm.ends_with(".fcs") {
+        println!("decrypting {} ...\n", input_path);
 
         let salt_len = 32;
         let nonce_len = 24;
-        let encrypted_file: Vec<u8> = get_file_as_byte_vec(&encrypted_file_path_norm)?;
+        let key_hash_ref_len = 32;
+        let encrypted_file: Vec<u8> = get_file_as_byte_vec(&input_path_norm)?;
 
         // Split the salt, nonce and the encrypted file
-        let (salt, data_file) = encrypted_file.split_at(salt_len);
-        let (nonce, ciphertext) = data_file.split_at(nonce_len);
+        let (salt, rem_data) = encrypted_file.split_at(salt_len);
+        let (nonce, rem_data) = rem_data.split_at(nonce_len);
+        let (key_hash_ref, ciphertext) = rem_data.split_at(key_hash_ref_len);
 
         let argon2_config = argon2_config();
-        let mut symmetric_key = argon2::hash_raw(passphrase.as_bytes(), salt, &argon2_config)?;
-        let cipher = XChaCha20Poly1305::new(symmetric_key[..32].as_ref().into());
-        let file_decrypted = cipher.decrypt(nonce.as_ref().into(), ciphertext.as_ref())?;
-        let file_stem_decrypted = &get_file_stem_to_string(&encrypted_file_path_norm)?;
-        let decrypted_file_path: String = format!("{}{}.zip", dest_dir_path_norm, file_stem_decrypted);
+        let mut key = argon2::hash_raw(passphrase.as_bytes(), salt, &argon2_config)?;
 
-        File::create(&decrypted_file_path)?;
-        fs::write(&decrypted_file_path, file_decrypted)?;
-        archiver::unarchive(&decrypted_file_path, &dest_dir_path_norm)?;
-        fs::remove_file(&decrypted_file_path)?;
-        println!("\ndecrypted to {}", dest_dir_path_norm);
+        // Hash the encryption key for comparison and compare it in constant time with the ref key hash
+        let key_hash: [u8; 32] = sha3_hash(&key)?;
+        let key_correct = constant_time_compare_256_bit(&key_hash, key_hash_ref[..32].try_into()?);
 
-        symmetric_key.zeroize();
-        passphrase.zeroize();
+        if key_correct {
+            let cipher = XChaCha20Poly1305::new(key[..32].as_ref().into());
+            let file_decrypted = cipher.decrypt(nonce.as_ref().into(), ciphertext.as_ref())?;
+            let file_stem_decrypted = &get_file_stem_to_string(&input_path_norm)?;
+            let decrypted_file_path: String = format!("{}{}.zip", output_dir_norm, file_stem_decrypted);
+
+            File::create(&decrypted_file_path)?;
+            fs::write(&decrypted_file_path, file_decrypted)?;
+            archiver::unarchive(&decrypted_file_path, &output_dir_norm)?;
+            fs::remove_file(&decrypted_file_path)?;
+            println!("\ndecrypted to {}", output_dir_norm);
+
+            key.zeroize();
+            passphrase.zeroize();
+        } else {
+            return Err(Message("The provided password is incorrect!".to_string()));
+        }
     } else {
         return Err(Message("This file should have '.fcs' extension!".to_string()));
     }
@@ -148,10 +188,10 @@ pub fn decrypt_file(encrypted_file_path: &str, dest_dir_path: &str, passphrase: 
 }
 
 // Encrypt large file, that doesn't fit in RAM, with XChaCha20Poly1305 algorithm. This is much slower
-pub fn encrypt_large_file(src_file_path: &str, dest_dir_path: &str, passphrase: &mut str) -> Result<(), CryptoError> {
-    let (src_file_path_norm, dest_dir_path_norm) = normalize_paths(src_file_path, dest_dir_path);
-    let file_stem = &archiver::archive(&src_file_path_norm, &dest_dir_path_norm)?;
-    let file_name_zipped = &format!("{dest_dir_path_norm}{file_stem}.zip");
+pub fn encrypt_large_file(input_path: &str, output_dir: &str, passphrase: &mut str) -> Result<(), CryptoError> {
+    let (input_path_norm, output_dir_norm) = normalize_paths(input_path, output_dir);
+    let file_stem = &archiver::archive(&input_path_norm, &output_dir_norm)?;
+    let file_name_zipped = &format!("{}{}.zip", output_dir_norm, file_stem);
     println!("\nencrypting {} ...", file_name_zipped);
 
     let mut source_file = File::open(file_name_zipped)?;
@@ -165,6 +205,9 @@ pub fn encrypt_large_file(src_file_path: &str, dest_dir_path: &str, passphrase: 
     let aead = XChaCha20Poly1305::new(key[..32].as_ref().into());
     let mut stream_encryptor = stream::EncryptorBE32::from_aead(aead, nonce.as_ref().into());
 
+    // Hash the encryption key for comparison when decrypting
+    let key_hash_ref: [u8; 32] = sha3_hash(&key)?;
+
     // XChaCha20-Poly1305 is an AEAD cipher and appends a 16 bytes authentication tag to each encrypted message, so the buffer becomes 516 bits
     const BUFFER_LEN: usize = 500;
     let mut buffer = [0u8; BUFFER_LEN];
@@ -172,10 +215,11 @@ pub fn encrypt_large_file(src_file_path: &str, dest_dir_path: &str, passphrase: 
         .write(true)
         .append(true)
         .create_new(true)
-        .open(format!("{}{}.fcls", &dest_dir_path_norm, file_stem))?;
+        .open(format!("{}{}.fcls", &output_dir_norm, file_stem))?;
 
     file_path_encrypted.write_all(&salt)?;
     file_path_encrypted.write_all(&nonce)?;
+    file_path_encrypted.write_all(&key_hash_ref)?;
 
     loop {
         let read_count = source_file.read(&mut buffer)?;
@@ -195,9 +239,9 @@ pub fn encrypt_large_file(src_file_path: &str, dest_dir_path: &str, passphrase: 
     }
 
     fs::remove_file(file_name_zipped)?;
-    let file_name_encrypted = &format!("{dest_dir_path_norm}{file_stem}.fcls");
+    let file_name_encrypted = &format!("{}{}.fcls", output_dir_norm, file_stem);
     println!();
-    println!("encrypted to {file_name_encrypted}");
+    println!("encrypted to {}", file_name_encrypted);
 
     salt.zeroize();
     nonce.zeroize();
@@ -208,24 +252,18 @@ pub fn encrypt_large_file(src_file_path: &str, dest_dir_path: &str, passphrase: 
 }
 
 // Decrypt large file, that doesn't fit in RAM, with XChaCha20Poly1305 algorithm. This is much slower
-pub fn decrypt_large_file(encrypted_file_path: &str, dest_dir_path: &str, passphrase: &mut str) -> Result<(), CryptoError> {
-    let (encrypted_file_path_norm, dest_dir_path_norm) = normalize_paths(encrypted_file_path, dest_dir_path);
+pub fn decrypt_large_file(input_path: &str, output_dir: &str, passphrase: &mut str) -> Result<(), CryptoError> {
+    let (input_path_norm, output_dir_norm) = normalize_paths(input_path, output_dir);
 
-    if encrypted_file_path_norm.ends_with(".fcls") {
-        println!("decrypting {} ...\n", encrypted_file_path);
+    if input_path_norm.ends_with(".fcls") {
+        println!("decrypting {} ...\n", input_path);
 
         let mut salt = [0u8; 32];
         let mut nonce = [0u8; 19];
-        let mut encrypted_file = File::open(&encrypted_file_path_norm)?;
-        let file_stem_decrypted = &get_file_stem_to_string(&encrypted_file_path_norm)?;
-        let decrypted_file_path = format!("{}{}.zip", dest_dir_path_norm, file_stem_decrypted);
-        let mut decrypted_file = OpenOptions::new()
-            .write(true)
-            .append(true)
-            .create_new(true)
-            .open(&decrypted_file_path)?;
-
+        let mut key_hash_ref = [0u8; 32];
+        let mut encrypted_file = File::open(&input_path_norm)?;
         let mut read_count = encrypted_file.read(&mut salt)?;
+
         if read_count != salt.len() {
             return Err(Message("Error reading salt!".to_string()));
         }
@@ -235,42 +273,63 @@ pub fn decrypt_large_file(encrypted_file_path: &str, dest_dir_path: &str, passph
             return Err(Message("Error reading nonce!".to_string()));
         }
 
+        read_count = encrypted_file.read(&mut key_hash_ref)?;
+        if read_count != key_hash_ref.len() {
+            return Err(Message("Error reading key_hash_ref!".to_string()));
+        }
+
         let argon2_config = argon2_config();
         let mut key = argon2::hash_raw(passphrase.as_bytes(), &salt, &argon2_config)?;
         let aead = XChaCha20Poly1305::new(key[..32].as_ref().into());
         let mut stream_decryptor = stream::DecryptorBE32::from_aead(aead, nonce.as_ref().into());
 
-        // 500 bytes for the encrypted piece of data, and 16 bytes for the authentication tag, which was added on encryption
-        const BUFFER_LEN: usize = 500 + 16;
-        let mut buffer = [0u8; BUFFER_LEN];
+        // Hash the encryption key for comparison and compare it in constant time with the ref key hash
+        let key_hash: [u8; 32] = sha3_hash(&key)?;
+        let key_correct = constant_time_compare_256_bit(&key_hash, key_hash_ref[..32].try_into()?);
 
-        loop {
-            let read_count = encrypted_file.read(&mut buffer)?;
+        if key_correct {
+            let file_stem_decrypted = &get_file_stem_to_string(&input_path_norm)?;
+            let decrypted_file_path = format!("{}{}.zip", output_dir_norm, file_stem_decrypted);
+            let mut decrypted_file = OpenOptions::new()
+                .write(true)
+                .append(true)
+                .create_new(true)
+                .open(&decrypted_file_path)?;
 
-            if read_count == BUFFER_LEN {
-                let plaintext = stream_decryptor
-                    .decrypt_next(buffer.as_slice())
-                    .map_err(ChaCha20Poly1305Error)?;
-                decrypted_file.write_all(&plaintext)?;
-            } else if read_count == 0 {
-                break;
-            } else {
-                let plaintext = stream_decryptor
-                    .decrypt_last(&buffer[..read_count])
-                    .map_err(ChaCha20Poly1305Error)?;
-                decrypted_file.write_all(&plaintext)?;
-                break;
+            // 500 bytes for the encrypted piece of data, and 16 bytes for the authentication tag, which was added on encryption
+            const BUFFER_LEN: usize = 500 + 16;
+            let mut buffer = [0u8; BUFFER_LEN];
+
+            loop {
+                let read_count = encrypted_file.read(&mut buffer)?;
+
+                if read_count == BUFFER_LEN {
+                    let plaintext = stream_decryptor
+                        .decrypt_next(buffer.as_slice())
+                        .map_err(ChaCha20Poly1305Error)?;
+                    decrypted_file.write_all(&plaintext)?;
+                } else if read_count == 0 {
+                    break;
+                } else {
+                    let plaintext = stream_decryptor
+                        .decrypt_last(&buffer[..read_count])
+                        .map_err(ChaCha20Poly1305Error)?;
+                    decrypted_file.write_all(&plaintext)?;
+                    break;
+                }
             }
+
+            archiver::unarchive(&decrypted_file_path, &output_dir_norm)?;
+            fs::remove_file(&decrypted_file_path)?;
+            println!("\ndecrypted to {}", output_dir_norm);
+
+            salt.zeroize();
+            nonce.zeroize();
+            key.zeroize();
+            passphrase.zeroize();
+        } else {
+            return Err(Message("The provided password is incorrect!".to_string()));
         }
-
-        archiver::unarchive(&decrypted_file_path, &dest_dir_path_norm)?;
-        fs::remove_file(&decrypted_file_path)?;
-        println!("\ndecrypted to {dest_dir_path_norm}");
-
-        salt.zeroize();
-        nonce.zeroize();
-        key.zeroize();
-        passphrase.zeroize();
     } else {
         return Err(Message("This file should have '.fcls' extension!".to_string()));
     }
